devices: vfio_pci: PV handling for intel-lpss devices
When compiled with the direct feature, crosvm allows an intel-lpss
vfio-pci device to issue PV operations from the guest. This is currently
done via a BAR write to an unused region, offset [0x400, 0x800) in BAR0.

Remove this PV region from an intel-lpss device's BAR0 mmap, and
special-case it in write_bar(). In this patch it does nothing more than
print a warning.

This is only a temporary solution. A proper communication channel should
be implemented in the future.

BUG=b:232887201
TEST=boot manatee

Change-Id: Ibecee7ac1c54d5ccf43f2db893505854697bb188
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3764938
Commit-Queue: Tomasz Nowicki <tnowicki@google.com>
Tested-by: Peter Fang <peter.fang@intel.corp-partner.google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>