devices: vhost: user: pass vhost-user platform ops as a trait object
The platform ops are not used frequently and can be passed as a trait
object without a performance hit. Doing so also simplifies some of our
interfaces, i.e. DeviceRequestHandler and VsockBackend do not take a
generic argument anymore.

This is also helpful in order to bring the DeviceRequestHandler and
VsockBackend closer to each other in the prospect of unifying them.

BUG=b:217480043
TEST=send data between the host and a VM connected to a vhost-user vsock device using socat.
TEST=send data between a VVU vsock device inside a device VM and another guest using socat.

Change-Id: I84fe2d4ff5ca714d6efbde955e221b4dff0bc812
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4222841
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>