vm_control: use HotplugCommand for hotplug devices
Currently we use VfioCommand to hotplug only vfio devices, but
pcie switch will also be hotplugged into the system and other type
of devices could also be added later. So we change VfioCommand to
HotplugCommand and introduce HotplugDeviceInfo structure to track
details of hotplug devices and send it in HotplugCommand.

BUG=b:199986018
TEST=tools/presubmit

Change-Id: I1a7fb316178ffff002a65b4f57f6e857a0993caa
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3769816
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>