virtio: video: decoder: ffmpeg: make input buffers manage their event
Make the InputBuffer type able to send the event signaling its
availability to be reused itself, when it is dropped. This does not
bring any immediate benefit, but is going to be a required design for
the next CL, where we transfer its ownership to libavcodec in order to
avoid copying its content.

BUG=b:169295147
TEST=v4l2r's simple_decoder can decode a H.264 stream.

Change-Id: I822bac85d9cd5f87e450e4c7ec009bb75655e91f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3782036
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>