virtio: video: decoder: ffmpeg: use zero-copy for input buffers
Use the newly introduced AvBufferRef of our ffmpeg crate to give
transfer ownership of the input mapping to libavcodec and avoid a copy
of the input data. libavcodec will drop our input buffer when it is done
with it, which will make it automatically send the
NotifyEndOfInputBuffer to signal it can be reused.

BUG=b:169295147
TEST=v4l2r's simple_decoder can decode a H.264 stream.

Change-Id: I9322e5b236a511ba3b8f064080111cfb3133d0e5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3782038
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>