main: convert gpu options to use serde_keyvalue
Convert the --gpu and --gpu-display parameters to use serde_keyvalue
instead of our own parser. Since the GPU display options can also be
specified along with the regular GPU options, this requires us to keep a
dedicated parsing function to maintain compatibility, but we can still
use the serde_keyvalue parsing within it.

Also add a few extra tests to ensure we didn't break anything in the
process.

This results in a net decrease of the ChromeOS crosvm binary size of
8KB.

BUG=b:218223240
TEST=cargo test --features "gpu" parse_gpu

Change-Id: Ia9daecb1d432e17b551bf36bfa0674e7df8459f1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3784660
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>