devices: vvu: proxy: update error message in create_sibling_guest_memory
Currently, vvu on crosvm does not allow the sibling VM to have the same
or more memory than the device VM. However, the memory size is not checked
in vvu proxy, resulting in an unclear error message generated when
attempting to map the memory.

Fix this by generating more relevent error when the memory size of
sibling VM is larger than the allocated memory size of device VM.

BUG=b:240498975
TEST=Start sibling VM with larger memory size than the device VM and
check if the generated error message.

Change-Id: I7cf25a0ba2e8798d5fab7c5bed88b97fe8428890
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3791936
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Morg <morg@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@google.com>
Tested-by: Keita Suzuki <suzukikeita@google.com>