gpu: support display mode selection
We will support two display modes on Windows:

(1) Borderless full screen mode, which can be specified with
    "--gpu-display borderless_full_screen".
(2) Windowed mode, which can be specified with the expected display
    size, e.g. "--gpu-display width=800,height=600".

If none of them is specified, we use windowed mode by default. Note
that on Unix, only windowed mode is supported for now.

BUG=b:233676779
TEST=presubmit

Change-Id: I5665ceedb31533c7c298eb7aa50e23eafb1e9411
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3806755
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Pujun Lun <lunpujun@google.com>
Commit-Queue: Pujun Lun <lunpujun@google.com>
Reviewed-by: Kaiyi Li <kaiyili@google.com>