base: use tempfile in syslog_file test
The test isn't unix specific anymore, so we ought to move it to
base/src/syslog.rs, however, when moved, it fails with an "Illegal Seek"
error when run in the presubmit (but not in `cargo test`). The cause is
unclear.

BUG=b:228881485
TEST=./tools/dev_container ./tools/presubmit

Change-Id: I76e6595007a3b59f6bbd5a4afae42f911686f3e3
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3817974
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Auto-Submit: Frederick Mayle <fmayle@google.com>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>