devices: pflash: don't set status on each read
QEMU's probe of the pflash device[1] expects to be able to clear the status
register, perform a read of the same location, and then read the status
register with the cleared state. To accomplish this only set the READY
status on an explicit READ_ARRAY command.

[1]: https://github.com/tianocore/edk2/blob/master/OvmfPkg/QemuFlashFvbServicesRuntimeDxe/QemuFlash.c

BUG=b:238344840
TEST=Added test for specific case

Change-Id: I81b326d4a546bb03fb55b766b630153ffbc758c5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3831261
Tested-by: Mike Gerow <gerow@google.com>
Auto-Submit: Mike Gerow <gerow@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>