infra: Fix coverage uploads
codecov tries to look for coverage files but fails if there are
inaccessible directories (which we occasionally have when they
are created by docker).
This change disables the search, which is fine since we already
give it the name of the file to upload.

BUG=b:239255082
TEST=led get-builder luci.crosvm.ci:linux_x86_64
  | led edit-recipe-bundle
  | led launch

Change-Id: Ie18f5e01dae486db45ba3434813a0799b99c2f1c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3842810
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Tested-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>